:recycle: change EditorJs component style and move location
