+++ src/components/editorjs/i18n.ts
